Linux 32bit build fixed.
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk@3817 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ pwsafe/pwsafe/src/core/PWCharPool.cpp
@@ -412,7 +412,7 @@ StringX CPasswordCharPool::MakePronounceable() const
 
 bool CPasswordCharPool::CheckPassword(const StringX &pwd, StringX &error)
 {
-  const int MinLength = 8;
+  const size_t MinLength = 8;
   size_t length = pwd.length();
   // check for minimun length
   if (length < MinLength) {
